US121271 prep - move summarizer-styles into activity-editor base folder
