+++ components/d2l-activity-editor/activity-summarizer-styles.js
